% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_lasso.R
\name{linear}
\alias{linear}
\title{linear}
\usage{
linear(X, y)
}
\arguments{
\item{X}{independent variable(s)}

\item{y}{dependent variable}
}
\value{

}
\description{
linear is used to fit linear models
}
