% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_lasso.R
\name{logistic}
\alias{logistic}
\title{logistic}
\usage{
logistic(X, y)
}
\arguments{
\item{X}{independent variable(s)}

\item{y}{dependent variable}
}
\value{

}
\description{
logistic is used to fit logistic models
}
