% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_lasso.R
\name{random_lasso}
\alias{random_lasso}
\title{random lasso}
\usage{
random_lasso(X, y, B, q1, q2, continuous)
}
\arguments{
\item{X}{independent variables.}

\item{y}{dependent variable.}

\item{B}{bootstrap samples.}

\item{q1}{q1 candidate variables.}

\item{q2}{q2 candidate variables.}

\item{continuous}{whether the dependent variable is continuous. Default is true.}
}
\value{

}
\description{
random lasso is used to fit a generalized linear model by random lasso regression.
}
