% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_lasso.R
\name{ridge}
\alias{ridge}
\title{ridge}
\usage{
ridge(X, y, continuous = TRUE, exclude = NULL, intercept = TRUE)
}
\arguments{
\item{X}{independent variable}

\item{y}{dependent variable}

\item{continuous}{whether the dependent variable is continuous. Default is true.}

\item{exclude}{Indices of variables to be excluded from the model. Default is none.}

\item{intercept}{Should intercept(s) be fitted (default=TRUE) or set to zero (FALSE).}
}
\value{

}
\description{
Fit a generalized linear model by ridge regression.
Can deal with continuous and binary dependent variable Y.
}
\examples{
# continuous
ridge(X = mtcars[,-1], y = mtcars[,1])

# binary
rigde(X=dat[,-1],y=dat[,1],continuous=F)
}
